DesktopSearchEngine/requirements.txt
PyPDF2 == 2.5.0
python-docx == 0.8.11
google-api-python-client==1.6.4
httplib2==0.10.3
oauth2client==4.1.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
rsa==3.4.2
six==1.12.0
SpeechRecognition==3.8.1
tqdm==4.19.5
uritemplate==3.0.0